+++ src/output/transform.rs
@@ -7,7 +7,7 @@
 
 use super::cssbuf::{CssBuf, CssHead};
 use crate::css::{BodyItem, Comment, Import, Property, Rule, Selectors};
-use crate::error::Error;
+use crate::error::{Error, Invalid};
 use crate::file_context::FileContext;
 use crate::parser::Parsed;
 use crate::sass::{
@@ -262,9 +262,10 @@ fn handle_item(
             buf.join(sub);
         }
         Item::AtRule {
-            ref name,
-            ref args,
-            ref body,
+            name,
+            args,
+            body,
+            pos: _,
         } => {
             buf.do_separate();
             buf.do_indent_no_nl();
@@ -319,8 +320,9 @@ fn handle_item(
             {
                 // Ok, this is cheating for the test suite ...
                 let p = pos.clone().opt_back("@function ");
-                return Err(Error::InvalidFunctionName(p));
+                return Err(Invalid::FunctionName.at(p));
             }
+            check_body(body, BodyContext::Function)?;
             scope.define_function(
                 name.into(),
                 Function::closure(
@@ -331,13 +333,12 @@ fn handle_item(
                 ),
             );
         }
-        Item::Return(_) => {
-            return Err(Error::S(
-                "Return not allowed in plain context".into(),
-            ));
+        Item::Return(_, ref pos) => {
+            return Err(Invalid::AtRule.at(pos.clone()));
         }
 
         Item::MixinDeclaration(ref name, ref args, ref body, ref pos) => {
+            check_body(body, BodyContext::Mixin)?;
             scope.define_mixin(
                 name.into(),
                 MixinDecl::new(args, &scope, body, pos),
@@ -362,14 +363,7 @@ fn handle_item(
                     file_context,
                 )
                 .map_err(|e: Error| match e {
-                    Error::BadArguments(msg, decl) => {
-                        let pos = pos.in_call(name);
-                        Error::BadCall(msg, pos, Some(decl))
-                    }
-                    Error::AtError(msg, _pos) => {
-                        let msg = format!("Error: {}", msg);
-                        Error::BadCall(msg, pos.clone(), None)
-                    }
+                    Error::Invalid(err, _) => err.at(pos.clone()),
                     e => {
                         let pos = pos.in_call(name);
                         Error::BadCall(e.to_string(), pos, None)
@@ -410,9 +404,11 @@ fn handle_item(
         Item::IfStatement(ref cond, ref do_if, ref do_else) => {
             let cond = cond.evaluate(scope.clone())?.is_true();
             let items = if cond { do_if } else { do_else };
+            check_body(items, BodyContext::Control)?;
             handle_body(items, head, rule, buf, scope, file_context)?;
         }
         Item::Each(ref names, ref values, ref body) => {
+            check_body(body, BodyContext::Control)?;
             let mut rule = rule;
             let pushed = scope.store_local_values(names);
             for value in values.evaluate(scope.clone())?.iter_items() {
@@ -440,6 +436,7 @@ fn handle_item(
                 to.evaluate(scope.clone())?,
                 *inclusive,
             )?;
+            check_body(body, BodyContext::Control)?;
             let mut rule = rule;
             for value in range {
                 let scope = ScopeRef::sub(scope.clone());
@@ -455,6 +452,7 @@ fn handle_item(
             }
         }
         Item::While(ref cond, ref body) => {
+            check_body(body, BodyContext::Control)?;
             let mut rule = rule;
             let scope = ScopeRef::sub(scope);
             while cond.evaluate(scope.clone())?.is_true() {
@@ -476,13 +474,14 @@ fn handle_item(
             eprintln!("WARNING: {}", value.evaluate(scope)?.format(format));
         }
         Item::Error(ref value, ref pos) => {
-            return Err(Error::AtError(
+            return Err(Invalid::AtError(
                 value.evaluate(scope)?.format(format).to_string(),
-                pos.clone(),
-            ));
+            )
+            .at(pos.clone()));
         }
 
         Item::Rule(ref selectors, ref body) => {
+            check_body(body, BodyContext::Rule)?;
             if rule.is_none() {
                 buf.do_separate();
             }
@@ -530,6 +529,7 @@ fn handle_item(
         }
         Item::NamespaceRule(ref name, ref value, ref body) => {
             if let Some(rule) = rule {
+                check_body(body, BodyContext::NsRule)?;
                 let value = value.evaluate(scope.clone())?;
                 let name = name.evaluate(scope.clone())?;
                 if !value.is_null() {
@@ -581,3 +581,93 @@ fn handle_item(
     }
     Ok(())
 }
+
+#[derive(Clone, Copy, PartialEq, Eq)]
+enum BodyContext {
+    Mixin,
+    Function,
+    Control,
+    Rule,
+    NsRule,
+}
+
+fn check_body(body: &[Item], context: BodyContext) -> Result<(), Error> {
+    for item in body {
+        match item {
+            Item::Forward(_, _, _, _, pos) => {
+                return Err(Invalid::AtRule.at(pos.clone()));
+            }
+            Item::Use(_, _, _, pos) => {
+                return Err(Invalid::AtRule.at(pos.clone()));
+            }
+            Item::MixinDeclaration(.., ref pos) => {
+                let pos = pos.clone().opt_back("@mixin ");
+                match context {
+                    BodyContext::Mixin => {
+                        return Err(Invalid::MixinInMixin.at(pos));
+                    }
+                    BodyContext::Control => {
+                        return Err(Invalid::MixinInControl.at(pos));
+                    }
+                    BodyContext::Rule => (), // This is ok
+                    _ => {
+                        return Err(Invalid::AtRule.at(pos.opt_trail_ws()));
+                    }
+                }
+            }
+            Item::FunctionDeclaration(_, _, ref pos, _) => {
+                let pos = pos.clone().opt_back("@function ");
+                match context {
+                    BodyContext::Mixin => {
+                        return Err(Invalid::FunctionInMixin.at(pos));
+                    }
+                    BodyContext::Control => {
+                        return Err(Invalid::FunctionInControl.at(pos));
+                    }
+                    BodyContext::Rule => (), // This is ok
+                    _ => {
+                        return Err(Invalid::AtRule.at(pos.opt_trail_ws()));
+                    }
+                }
+            }
+            Item::Return(_, ref pos) if context != BodyContext::Function => {
+                return Err(Invalid::AtRule.at(pos.clone()));
+            }
+            Item::AtRule {
+                name,
+                args: _,
+                body: _,
+                pos,
+            } if context != BodyContext::Rule => {
+                if !name
+                    .single_raw()
+                    .map(|name| CSS_AT_RULES.contains(&name))
+                    .unwrap_or(false)
+                {
+                    return Err(Invalid::AtRule.at(pos.clone()));
+                }
+            }
+            _ => (),
+        }
+    }
+    Ok(())
+}
+
+const CSS_AT_RULES: [&str; 16] = [
+    "charset",
+    "color-profile",
+    "counter-style",
+    "document",
+    "font-face",
+    "font-feature-values",
+    "import",
+    "keyframes",
+    "layer",
+    "media",
+    "namespace",
+    "page",
+    "property",
+    "scroll-timeline",
+    "supports",
+    "viewport",
+];